Add exception for field searches specified by a string value
If a field search is specified by a symbol, the table name will be added
to the column in the generated SQL clauses. However, if the field search
is specified by a string, the table name will not be added.

This allows for the ability to specify complex field searches, such as
when searching a non-string column, e.g., CAST(foos.value AS text). The
table name must be included in the search string, when necessary, since
it will not be added automatically.